moved conversion functions to separate file as suggest by CL
